Change percentage values to be stored as numbers.
We were storing percentage values as strings with [num]%. The spec
says to store these as numbers instead. In the case of line, which
can be an integer or a percentage, we have to differentiate between
the two with the snap-to-lines flag. We have to expose this in the
WebIDL as well so we have to do the conversion at some point. I think
it's best to just store it as a number right off the bat.