Make parser able to handle tags with no end '>'.
We were failing to parse tags that have eol before an ending '>'.
According to spec in these circumstances we should be returning
what has been parsed in the tag up until that point and treat
the eol exactly like if we hit a '>'.

I've added some extra tests to check for this in the future on
tags that accept annotations.