Fix no spaces between cue times separator bug.
The problem was that after we consume a timestamp we would replace
everything that isn't a whitespace until the next whitespace. In
situations with no white spaces this means it would remove the
rest of the cue time data. Changed so that we replace up until
the next whitespace or "-" character.