Change textContent to innerHTML to support IE8.
The tests are breaking for me so I can't tell if this is a breaking change, but it looked like it might not be. textContent works like innerHTML execept it basically strips out the tags. It looke like the nodes were already text nodes, so this shouldn't be an issue.