Reverse step direction at the correct time.
The direction of the steps should only be reversed when
the axis along which the overlap between the container box
and the cue box is happening is the axis that the steps are
moving towards.

I've had to turn off the large-integer-value processing test
because it's hanging for a long time on the step algorithm.