Don't build the dist files unless we explicitly want to.
In the normal course of development we should instead rely on a dev
build of vtt.js. This minimizes the possibility of accidentally
building and commiting an official dist while developing.

I've also updated cue2json.js to use the dev build instead of node-vtt's
default page, which will be running an older version of vtt.js.