Fix issues with large line integer optimization.
A few things were going wrong:
1. I forgot to turn the test back on.
2. The previous optimization solution worked with most integers, however,
when the integer became to large it messed up due to JavaScripts way
of rounding floating point numbers over 64 bits.

This solution gets around the second issue by optimizing the amount it will
be moved to _before_ we have to do any real math on the line value,
whose value may be over 64 bits. However, we now have to disable the
'large-integer-value' test for another reason, which is known, having to
do with the computed value balancer messing up when the step direction
is flipped.